fix: Convert PositionFloat from decimal to double to resolve SQLite ordering issue (#85)
- Changed PlaylistTrack.PositionFloat from decimal to double type
- Removed precision configuration from database context
- Updated all DTOs and request models to use double
- Fixed all tests to use double literals instead of decimal
- Created migration to update database schema
- Updated API documentation to reflect type change

This resolves issue #84 where SQLite was unable to properly order by decimal
values in ORDER BY clauses. The double type is properly supported by SQLite
and maintains the fractional positioning system for conflict-free reordering.

+++ src/Audiarr.Core/Entities/PlaylistTrack.cs
@@ -5,7 +5,7 @@ public class PlaylistTrack
     public required string PlaylistId { get; set; }
     public required string TrackId { get; set; }
     public int Position { get; set; }
-    public decimal PositionFloat { get; set; } = 0; // For conflict-free reordering
+    public double PositionFloat { get; set; } = 0; // For conflict-free reordering
     public DateTime AddedAt { get; set; } = DateTime.UtcNow;
     public string? AddedBy { get; set; } // Username who added the track (for future collaborative features)
